requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
flask-jwt-extended==4.6.0
werkzeug==3.0.3
pyjwt==2.8.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
